Finished FileProcessFactory to process known CHPP Xml parsed files. Changed some classes access modifier. Modified IFileProcessStrategy to drop the parameters parameter. Improved ManagerCompendium FileProcess Strategy to detect if the file corresponds to the app's user. Finished all FileProcess Strategies. Added centralized FileProcesser. Fixed ApplicationObjects container registration. Changed EventArguments on DownloadManger related objects. Improved process progress report in both DownloadManager and FileProcessManager. Renamed DownloadProgressChangedEventArgs to FileTaskProgressChangedEventArgs to reuse it on both managers. Modified some business objetcs to unify the Season data type and related changes to mappings and file parsers. Fixed an issue on Players Xml file parser where the ActionType wasn't being parsed correctly. Updated and created localized strings. Added new icon and images for FormGenericProgress. Changed FormUser Load event to Shown event.

+++ Hyperar.HattrickUltimate.BusinessObjects/App/League.cs
@@ -54,7 +54,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the current Season.
         /// </summary>
-        public int CurrentSeason { get; set; }
+        public short CurrentSeason { get; set; }
 
         /// <summary>
         /// Gets or sets the number of divisions.
@@ -84,7 +84,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the season offset.
         /// </summary>
-        public int SeasonOffset { get; set; }
+        public short SeasonOffset { get; set; }
 
         /// <summary>
         /// Gets or sets the Senior Series.
